Split progress reports into reports module
