+++ apps/backend/src/progress/reports/progressReviewSchedule.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert/strict";
 import test from "node:test";
-import { HttpError } from "../shared/errors";
+import { HttpError } from "../../shared/errors";
 import {
   loadUserProgressReviewScheduleInExecutor,
   parseProgressReviewScheduleInputFromRequest,
